requirements.txt
langchain
langchain_community
langchain_core
langchain_openai
python-dotenv
streamlit
chromadb
beautifulsoup4
crewai==0.27.2
pydantic
langchain-anthropic
pysqlite3-binary
duckduckgo-search
